+++ R/validate.R
@@ -54,3 +54,37 @@ setValidity(
     arkhe::check_class(object, cnd)
   }
 )
+
+setValidity(
+  Class = "TimeIntervals",
+  method = function(object) {
+    ## Get data
+    names <- object@.Id
+    start <- object@.Start
+    end <- object@.End
+    m <- length(start)
+
+    ## Validate
+    cnd <- list(
+      arkhe::validate(arkhe::assert_length(names, m)),
+      arkhe::validate(arkhe::assert_length(end, m)),
+      arkhe::validate(arkhe::assert_missing(names)),
+      arkhe::validate(arkhe::assert_missing(start)),
+      arkhe::validate(arkhe::assert_missing(end)),
+      arkhe::validate(assert_ordered(start, end))
+    )
+
+    ## Return conditions, if any
+    arkhe::check_class(object, cnd)
+  }
+)
+
+assert_ordered <- function(start, end) {
+  arg_start <- deparse(substitute(start))
+  arg_end <- deparse(substitute(end))
+  if (any(start > end)) {
+    msg <- sprintf("%s is younger than %s.", sQuote(arg_start), sQuote(arg_end))
+    stop(msg)
+  }
+  invisible(NULL)
+}